requirements.txt
numpy
scipy
pandas
pyyaml
torch
gpytorch
matplotlib
pytest
scikit-learn
sqlalchemy
psycopg2-binary
streamlit
